fix some config

+++ gft-test-service/src/main/java/com/gft/inditex/repository/PriceRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PriceRepository extends JpaRepository<PriceEntity, Long> {
+public interface PriceRepository extends JpaRepository<PriceEntity, Integer> {
 
     //TODO methods
 }
